Add Git mirroring service
This adds the m2a-git-mirror utility to the image which can be used to
maintain a cache of Git repositories to speed up checkouts in Jenkins jobs.

For this work, git-daemon needs to be running on the Jenkins host, which is
what is actually used to serve fetch requests from clients. The right way to
do this would be to run a separate container in a side-car configuration, but
that would require a lot of changes to the Jenkins deployment configuration.

Instead, we adopt the slight anti-pattern of running multiple services
(git-daemon and Jenkins) in a single container. Specifically, these services
are managed by Supervisor which in turn runs under Tini (like the base Jenkins)
image.