Set HOME for Supervisor managed processes
Assumably, as Supervisor runs as root, its $HOME is set to /root. However,
when a user is specified for a managed process, this $HOME variable is
inherited. This would mean that Jenkins jobs (run as the 'jenkins' user)
would try to read some global configuration from /root, which of course,
it didn't have permission to do so.

This makes sure that for processes where we set the user, we also set
a corresponding, appropriate value for $HOME.